+++ HISTORY.md
@@ -1,3 +1,9 @@
+1.0.1 / 2016-08-31
+==================
+
+  * Add data type validation for the array content
+  * Add range validation to the array in order to match the amount of replaceable characters in the string
+
 1.0.0 / 2016-08-30
 ==================